Move ep-dispatcher to just dispatcher
