+++ src/index.js
@@ -1,6 +1,6 @@
 export ActionPrototype from './lib/ActionPrototype'
 export  ExtraStorage from './lib/extra_storage'
-export Dispatcher from './lib/ep-dispatcher'
+export Dispatcher from './lib/dispatcher'
 export JsonFetcher from './lib/jsonFetcher'
 //export FeatureFlags from './lib/FeatureFlags' -- haven;t quite figured out the singleton pattern here
 export * from './testUtils'
