+++ src/lib/ActionPrototype.js
@@ -1,5 +1,5 @@
 import fetcher from '../lib/jsonFetcher'
-import AppDispatcher  from '../lib/ep-dispatcher'
+import AppDispatcher  from '../lib/dispatcher'
 import KeyMirror from 'keymirror'
 
 //TODOS:
